requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Django==4.1.6
django-redis==5.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.3
pytz==2022.7.1
redis==4.4.2
sqlparse==0.2.4
